feat: merge speed optimizations (#2)
* feat: write document ids directly into the target file without binary encoding

* fix: write document lengths directly into destination without binary encoding

* fix: reduced number of copies by writing directly to bufio

* fix: remove copies from hot collision path

* fix: improved hot path by writing entire structs directly

* fix: improved write speed by removing almost all data binary conversions

* fix: write directly to files on big-buffer optimization scenario

* fix: keep using bufio even when writing directly

* fix: use sequential advice for mmap load

* fix: loop unrolling for bitmap operation

* fix: runtime.GC on every benchmark iteration

* fix: restored writes without allocs

* fix: support vector as a receiver

* feat: support for simd operations

* fix: use directly vectors over data

* feat: moved engine to roaring 32

* feat: removed experimental simd

* feat: parallel creation of document ids and the rest of the file

* fix: use counter instead

* fix: do not use append on buffer

* feat: updated .gitignore

* fix: removed dead map

* fix: improved benchmark initialization

* fix: correct setup of merge benchmark

* fix: parallel heavy bench

* fix: skip empty tokens

* fix: support for correct type for document count

* fix: capture parentheses

* chore(release): 1.3.0

+++ query/searcher.go
@@ -25,9 +25,9 @@ func New(s *storage.Storage) (searcher *Searcher) {
 
 // Once a filtering and scoring are done, next step of a searching algorithm
 // Resolves the ctx to an actual idx slice
-func (s *Searcher) ResolveScores(ctx *QueryContext) (idxs []uint64) {
+func (s *Searcher) ResolveScores(ctx *QueryContext) (idxs []uint32) {
 	type scoreEntry struct {
-		docIdx uint64
+		docIdx uint32
 		score  float64
 	}
 
@@ -59,7 +59,7 @@ func (s *Searcher) ResolveScores(ctx *QueryContext) (idxs []uint64) {
 		},
 	)
 
-	idxs = make([]uint64, len(scores))
+	idxs = make([]uint32, len(scores))
 	for index := range scores {
 		idxs[index] = scores[index].docIdx
 	}
