+++ testsuite/query.go
@@ -12,7 +12,7 @@ import (
 // RunQuery filters then scores a query against s, returning the ranked doc
 // indices (best first) alongside the populated context so assertions can read
 // raw scores and the resolved bitmap.
-func RunQuery(q *query.SimpleQuery, s *storage.Storage) (idxs []uint64, ctx *query.QueryContext) {
+func RunQuery(q *query.SimpleQuery, s *storage.Storage) (idxs []uint32, ctx *query.QueryContext) {
 	searcher := query.New(s)
 	searcher.LevenshteinM = levenshtein.DefaultM
 	searcher.LevenshteinMaxK = levenshtein.DefaultK
@@ -25,19 +25,19 @@ func RunQuery(q *query.SimpleQuery, s *storage.Storage) (idxs []uint64, ctx *que
 
 // IndexOfDocument returns the internal index assigned to an external doc id after
 // the alphabetical sort performed by SortAndBuildFrom.
-func IndexOfDocument(s *storage.Storage, id string) (uint64, bool) {
+func IndexOfDocument(s *storage.Storage, id string) (uint32, bool) {
 	asBytes := []byte(id)
 	for i := range s.DocumentsIds {
 		d := &s.DocumentsIds[i]
 		if bytes.Equal(asBytes, d.Value.Bytes()) {
-			return uint64(i), true
+			return uint32(i), true
 		}
 	}
 	return 0, false
 }
 
 // ResolveDocumentIndexes maps a ranked slice of internal indices back to external ids.
-func ResolveDocumentIndexes(s *storage.Storage, idxs []uint64) []string {
+func ResolveDocumentIndexes(s *storage.Storage, idxs []uint32) []string {
 	out := make([]string, len(idxs))
 	for i, idx := range idxs {
 		out[i] = strings.Clone(s.DocumentsIds[idx].Value.UnsafeString())
@@ -49,13 +49,13 @@ func ResolveDocumentIndexes(s *storage.Storage, idxs []uint64) []string {
 // then resolves to a ranked slice (best first). Passing candidates == nil means
 // "every document in the corpus"; a non-nil (even empty) slice restricts the
 // candidate set to exactly those internal indices.
-func RunFieldScore(s *storage.Storage, fieldHash uint64, candidates []uint64) (idxs []uint64, ctx *query.QueryContext) {
+func RunFieldScore(s *storage.Storage, fieldHash uint64, candidates []uint32) (idxs []uint32, ctx *query.QueryContext) {
 	searcher := query.New(s)
 	ctx = &query.QueryContext{}
 
 	if candidates == nil {
 		for i := range s.DocumentsIds {
-			ctx.Bitmap.Add(uint64(i))
+			ctx.Bitmap.Add(uint32(i))
 		}
 	} else {
 		for _, c := range candidates {
